backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "apscheduler>=3.11.2",
    "fastapi[standard]>=0.136.0",
    "httpx>=0.28.1",
    "numpy>=2.4.4",
    "pillow>=12.2.0",
    "prometheus-client>=0.25.0",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.44.0",
]
